create service to resolve and cache entity set ids
