--- src/main/java/com/openlattice/chronicle/services/entitysets/EntitySetsManager.java
@@ -1,7 +1,0 @@
-package com.openlattice.chronicle.services.entitysets;
-
-/**
- * @author alfoncenzioka &lt;alfonce@openlattice.com&gt;
- */
-public interface EntitySetsManager {
-}
